Added default admin security checks to all data access API calls.
git-svn-id: http://source.pentaho.org/svnroot/bi-platform-v2/trunk@37712 5d404267-db13-0410-8476-b1f6b398e63a

+++ bi-platform-data-access/src/org/pentaho/platform/dataaccess/datasource/wizard/service/impl/ConnectionServiceDelegate.java
@@ -1,9 +1,11 @@
 package org.pentaho.platform.dataaccess.datasource.wizard.service.impl;
 
+import java.lang.reflect.Constructor;
 import java.util.ArrayList;
 import java.util.List;
 
 import org.pentaho.commons.connection.IPentahoConnection;
+import org.pentaho.platform.api.engine.IPluginResourceLoader;
 import org.pentaho.platform.api.repository.datasource.DatasourceMgmtServiceException;
 import org.pentaho.platform.api.repository.datasource.IDatasource;
 import org.pentaho.platform.api.repository.datasource.IDatasourceMgmtService;
@@ -11,19 +13,50 @@ import org.pentaho.platform.dataaccess.datasource.IConnection;
 import org.pentaho.platform.dataaccess.datasource.wizard.service.ConnectionServiceException;
 import org.pentaho.platform.engine.core.system.PentahoSystem;
 import org.pentaho.platform.engine.services.connection.PentahoConnectionFactory;
+import org.pentaho.platform.plugin.services.webservices.SessionHandler;
 import org.pentaho.platform.repository.hibernate.HibernateUtil;
 
 public class ConnectionServiceDelegate {
+  
+  private IDataAccessPermissionHandler dataAccessPermHandler;
   private IDatasourceMgmtService datasourceMgmtSvc;
   
   public ConnectionServiceDelegate() {
-   }
+  }
 
+  protected boolean hasDataAccessPermission() {
+    if (dataAccessPermHandler == null) {
+      String dataAccessClassName = null;
+      try {
+        IPluginResourceLoader resLoader = PentahoSystem.get(IPluginResourceLoader.class, null);
+        dataAccessClassName = resLoader.getPluginSetting(getClass(), "settings/data-access-permission-handler", "org.pentaho.dataaccess.datasource.wizard.service.impl.SimpleDataAccessPermissionHandler" );  //$NON-NLS-1$ //$NON-NLS-2$
+        Class<?> clazz = Class.forName(dataAccessClassName);
+        Constructor<?> defaultConstructor = clazz.getConstructor(new Class[]{});
+        dataAccessPermHandler = (IDataAccessPermissionHandler)defaultConstructor.newInstance(new Object[]{});
+      } catch (Exception e) {
+        // TODO: error(Messages.getErrorString("DashboardRenderer.ERROR_0024_SQL_PERMISSIONS_INIT_ERROR", sqlExecClassName), e); //$NON-NLS-1$
+        e.printStackTrace();
+        
+        // TODO: Unhardcode once this is an actual plugin
+        dataAccessPermHandler = new SimpleDataAccessPermissionHandler();
+      }
+
+      
+    }
+    return dataAccessPermHandler != null && dataAccessPermHandler.hasDataAccessPermission(SessionHandler.getSession());
+  }
+  
   public ConnectionServiceDelegate(IDatasourceMgmtService datasourceMgmtSvc) {
     this.datasourceMgmtSvc = datasourceMgmtSvc;
   }
   
   public List<IConnection> getConnections() throws ConnectionServiceException  {
+    if (!hasDataAccessPermission()) {
+      // TODO: log
+      System.out.println("NO PERMISSION");
+      return null;
+    }
+    
     List<IConnection> connectionList = new ArrayList<IConnection>();
     try  {
       for(IDatasource datasource:datasourceMgmtSvc.getDatasources()) {
@@ -34,15 +67,27 @@ public class ConnectionServiceDelegate {
     }
     return connectionList;
   }
+
   public IConnection getConnectionByName(String name) throws ConnectionServiceException  {
+    if (!hasDataAccessPermission()) {
+      // TODO: log
+      System.out.println("NO PERMISSION");
+      return null;
+    }
     try {
       return convertTo(datasourceMgmtSvc.getDatasource(name));
     } catch(DatasourceMgmtServiceException dme) {
       throw new ConnectionServiceException("Unable to get connection: " + name + " " + dme.getLocalizedMessage(), dme);  
     }
-    
   }
+  
   public Boolean addConnection(IConnection connection) throws ConnectionServiceException  {
+    if (!hasDataAccessPermission()) {
+      // TODO: log
+      System.out.println("NO PERMISSION");
+      return null;
+    }
+
     try {
       HibernateUtil.beginTransaction();
       datasourceMgmtSvc.createDatasource(convertFrom(connection));
@@ -52,7 +97,13 @@ public class ConnectionServiceDelegate {
       throw new ConnectionServiceException("Unable to add connection: " + connection.getName() + " "  + e.getLocalizedMessage(), e);  
     }
   }
+
   public Boolean updateConnection(IConnection connection) throws ConnectionServiceException  {
+    if (!hasDataAccessPermission()) {
+      // TODO: log
+      System.out.println("NO PERMISSION");
+      return null;
+    }
     try {
       HibernateUtil.beginTransaction();
       datasourceMgmtSvc.updateDatasource(convertFrom(connection));
@@ -62,7 +113,13 @@ public class ConnectionServiceDelegate {
       throw new ConnectionServiceException("Unable to update connection: " + connection.getName() + " "  + e.getLocalizedMessage(),e);  
     }
   }
+  
   public Boolean deleteConnection(IConnection connection) throws ConnectionServiceException  {
+    if (!hasDataAccessPermission()) {
+      // TODO: log
+      System.out.println("NO PERMISSION");
+      return null;
+    }
     try {
       HibernateUtil.beginTransaction();
       datasourceMgmtSvc.deleteDatasource(convertFrom(connection));
@@ -73,6 +130,12 @@ public class ConnectionServiceDelegate {
     }
   }
   public Boolean deleteConnection(String name) throws ConnectionServiceException  {
+    if (!hasDataAccessPermission()) {
+      // TODO: log
+      System.out.println("NO PERMISSION");
+      return null;
+    }
+
     try {
       HibernateUtil.beginTransaction();
       datasourceMgmtSvc.deleteDatasource(name);
@@ -84,6 +147,11 @@ public class ConnectionServiceDelegate {
   }
   
   public boolean testConnection(IConnection connection) throws ConnectionServiceException {
+    if (!hasDataAccessPermission()) {
+      // TODO: log
+      System.out.println("NO PERMISSION");
+      return false;
+    }
     IPentahoConnection pentahoConnection = null;
       pentahoConnection = PentahoConnectionFactory.getConnection(IPentahoConnection.SQL_DATASOURCE, connection.getDriverClass(),
         connection.getUrl(), connection.getUsername(), connection.getPassword(), null,null);
